@@ -14,7 +14,6 @@ classifier =
         License :: OSI Approved :: MIT License
         Operating System :: POSIX :: Linux
         Programming Language :: Python :: 3
-        Programming Language :: Python :: 3.6
         Programming Language :: Python :: 3.7
         Programming Language :: Python :: 3.8
         Programming Language :: Python :: 3.9
@@ -32,7 +31,7 @@ install_requires =
 packages = find:
 package_dir =
     =src
-python_requires = >=3.6
+python_requires = >=3.7
 zip_safe = False
 
 [options.package_data]
