@@ -1,12 +1,12 @@
-from typing import Optional, Union
+from __future__ import annotations
 
 from .condatools import InCondaPkgCtx
 from .wheeltools import InWheelCtx
 
 
 def InGenericPkgCtx(
-    in_path: str, out_path: Optional[str] = None
-) -> Union[InWheelCtx, InCondaPkgCtx]:
+    in_path: str, out_path: str | None = None
+) -> InWheelCtx | InCondaPkgCtx:
     """Factory that returns a InWheelCtx or InCondaPkgCtx
     context manager depending on the file extension
     """
