@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 from typing import Dict, List, Set
 
@@ -6,9 +8,9 @@ from . import load_policies
 log = logging.getLogger(__name__)
 
 
-def versioned_symbols_policy(versioned_symbols: Dict[str, Set[str]]) -> int:
+def versioned_symbols_policy(versioned_symbols: dict[str, set[str]]) -> int:
     def policy_is_satisfied(
-        policy_name: str, policy_sym_vers: Dict[str, Set[str]]
+        policy_name: str, policy_sym_vers: dict[str, set[str]]
     ) -> bool:
         policy_satisfied = True
         for name in set(required_vers) & set(policy_sym_vers):
